backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.4
gunicorn==22.0.0
numpy>=1.26,<3
httpx==0.27.2
SQLAlchemy==2.0.36
psycopg[binary]==3.2.5
redis==5.2.1
python-jose[cryptography]==3.3.0
prometheus-client==0.21.1
